:sparkles: feat: trazendo os funcionarios para a lista de funcionarios no form de devolucao
