+++ devolucao/static/css/devolucao.css
@@ -1,6 +1,6 @@
 .oldest-item {
-            background-color: #fef2f2 !important;
-            border-color: #fecaca !important;
+    background-color: #fef2f2 !important;
+    border-color: #fecaca !important;
 }
 .oldest-item .text-danger {
     color: #dc2626 !important;
